+++ lib/adapty_flutter.dart
@@ -25,6 +25,7 @@ export 'src/models/adapty_subscription_period.dart' show AdaptySubscriptionPerio
 export 'src/models/adapty_subscription_phase.dart' show AdaptySubscriptionPhase;
 export 'src/models/adapty_subscription_offer.dart' show AdaptySubscriptionOffer;
 export 'src/models/adapty_subscription_offer_identifier.dart' show AdaptySubscriptionOfferIdentifier, AdaptySubscriptionOfferType;
+export 'src/models/adapty_purchase_parameters.dart' show AdaptyPurchaseParameters, AdaptyPurchaseParametersBuilder;
 export 'src/models/adapty_profile.dart' show AdaptyProfile;
 export 'src/models/adapty_profile_gender.dart' show AdaptyProfileGender;
 export 'src/models/adapty_profile_parameters.dart' show AdaptyProfileParameters;
@@ -36,6 +37,7 @@ export 'src/models/product_reference.dart' show ProductReference;
 export 'src/models/adapty_paywall_view_configuration.dart' show AdaptyPaywallViewConfiguration;
 export 'src/models/adapty_remote_config.dart' show AdaptyRemoteConfig;
 export 'src/models/adapty_refund_preference.dart' show AdaptyRefundPreference;
+export 'src/models/adapty_installation_details.dart' show AdaptyInstallationDetails, AdaptyInstallationStatus, AdaptyInstallationStatusNotAvailable, AdaptyInstallationStatusNotDetermined, AdaptyInstallationStatusDetermined;
 
 export 'src/models/custom_assets/adaptyui_custom_assets.dart' show AdaptyCustomAsset, AdaptyCustomAssetColor, AdaptyCustomAssetLinearGradient;