+++ lib/src/models/private/adapty_purchase_result_json_builder.dart
@@ -11,7 +11,12 @@ extension AdaptyPurchaseResultJSONBuilder on AdaptyPurchaseResult {
         return AdaptyPurchaseResultUserCancelled();
       case _TypeKeys.success:
         final profileMap = json.object(_PurchaseResultKeys.profile);
-        return AdaptyPurchaseResultSuccess._(AdaptyProfileJSONBuilder.fromJsonValue(profileMap));
+        final jwsTransaction = json.stringIfPresent(_PurchaseResultKeys.jwsTransaction);
+
+        return AdaptyPurchaseResultSuccess._(
+          AdaptyProfileJSONBuilder.fromJsonValue(profileMap),
+          jwsTransaction,
+        );
       default:
         throw ArgumentError.value(type, _PurchaseResultKeys.type, 'Invalid purchase result type');
     }
@@ -21,6 +26,7 @@ extension AdaptyPurchaseResultJSONBuilder on AdaptyPurchaseResult {
 class _PurchaseResultKeys {
   static const type = 'type';
   static const profile = 'profile';
+  static const jwsTransaction = 'jws_transaction';
 }
 
 class _TypeKeys {