+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 3.10.0
+- AdaptyUI: added support for purchase button animations
+- AdaptyUI: added support for `productPurchaseParams` in `.createPaywallView` method.
+
+**Deprecations:** 
+- In `.makePurchase` method instead of `subscriptionUpdateParams` and `isOfferPersonalized` arguments use `parameters`.
+- In `AdaptyPaywall` model instead of `vendorProductIds`  use `productIdentifiers`.
+
+# 3.9.0
+- Added support for the User Acquisition feature. [Read more](https://adapty.io/docs/user-acquisition).
+- Minimum required Xcode version is now 16.0.
+- Various bug fixes and improvements.
+
+# 3.8.4
+- Changed the `makePurchase()` method signature when using parameters other than `product` – they have been extracted into `AdaptyPurchaseParameters`
+- [Android] Migrated to PBL v7
+- [Android] Added support for `obfuscatedProfileId` and `obfuscatedAccountId`
+- [iOS] Fixed support for Xcode 26
+
 # 3.8.3
 - [Android] Fixed AdaptyUI rendering issues